package refactor & removed password from validation query
