+++ src/main/java/com/pretz/dyntreeserver/repository/UserRepo.java
@@ -1,12 +1,12 @@
-package com.pretz.dyntreeserver;
+package com.pretz.dyntreeserver.repository;
 
 import com.pretz.dyntreeserver.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, Long> {
 
-    Optional<User> findByNameAndPassword(String name, String password);
+    Optional<User> findByName(String name);
     Optional<User> findByEmail(String email);
 }
